fixup! fixup! Split-off vault hook from vault secret backend
